remove abs from QmuTactile.cpp, its causing build error

+++ QmuTactile.cpp
@@ -1,5 +1,6 @@
 #include "Arduino.h"
 #include "QmuTactile.h"
+#include <cstdlib>
 
 QmuTactile::QmuTactile(uint8_t pin) {
     _pin = pin;
@@ -17,7 +18,7 @@ void QmuTactile::loop(void) {
         _nextPressingEvent = 0;
     }
 
-    const uint32_t buttonTime = abs(millis() - _pressMillis);
+    const uint32_t buttonTime = abs(long(millis() - _pressMillis));
 
     //Pressing
     if (pinState == LOW && buttonTime > TACTILE_PRESSING_TIME) {
